pyproject.toml
[tool.poetry]
name = "repl_python3_AniML"
version = "0.1.0"
description = ""
authors = ["szymonkop <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-telegram-bot = "^13.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
